+++ src/run.py
@@ -1,5 +1,0 @@
-import lo2yw7
-
-
-def export_yw():
-    lo2yw7.export_yw()