+++ src/yw_cnv.py
@@ -1,0 +1,13 @@
+"""Starter script for the odt/ods to yw7 converter. 
+
+Version @release
+Requires Python 3.6+
+Copyright (c) 2023 Peter Triesberger
+For further information see https://github.com/peter88213/lo2yw7
+Published under the MIT License (https://opensource.org/licenses/mit-license.php)
+"""
+import lo2yw7
+
+
+def export_yw():
+    lo2yw7.export_yw(XSCRIPTCONTEXT)