+++ docs/changelog.md
@@ -4,6 +4,12 @@
 
 ## Changelog
 
+### v1.2.3
+
+- Speed up execution by using a starter script.
+
+Based on PyWriter v12.1.0
+
 ### v1.2.2
 
 - When exporting to yw7, save the document only if modified.